Adding test cases for more than a billion

+++ src/main/java/NumberConvertor/INumberConvertor.java
@@ -1,5 +1,5 @@
 package NumberConvertor;
 
 public interface INumberConvertor {
-    String convertNumberToWords(int number);
+    String convertNumberToWords(long number);
 }
